blog-graphql-api/requirements.txt
aiodns==1.1.1
aiohttp==3.4.4
aniso8601==3.0.2
astroid==2.0.4
async-timeout==3.0.0
attrs==18.2.0
autopep8==1.4
cchardet==2.1.1
chardet==3.0.4
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
idna==2.7
idna-ssl==1.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
motor==2.0.0
motorodm==0.0.1
multidict==4.4.0
pkg-resources==0.0.0
promise==2.1
pycares==2.3.0
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pymongo==3.7.1
rope==0.11.0
Rx==1.6.1
six==1.11.0
typed-ast==1.1.0
typing==3.6.6
wrapt==1.10.11
yarl==1.2.6
